feat: 일부 user_id 체크없는 곳 추가

+++ src/pages/api/todo/deleteToDo.ts
@@ -5,9 +5,9 @@ export default async function deleteToDo(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
-  console.log("body.data", req.body.data);
+  console.log("delete todo body.data", req.body);
   db.query(
-    `DELETE FROM op_todo WHERE no IN (${req.body.data})`,
+    `DELETE FROM op_todo WHERE no IN (${req.body.data}) AND user_id="${req.body.user_id}"`,
     function (err: any, result: any) {
       if (err) {
         // console.log("에러는 ?", err);
